complete logging and tests
