--- python_modules/libraries/airflow-op-to-dagster-op/airflow_op_to_dagster_op/version.py
@@ -1,1 +1,0 @@
-__version__ = "dev"
