--- python_modules/libraries/airflow-op-to-dagster-op/tox.ini
@@ -1,30 +1,0 @@
-[tox]
-envlist = py{38,37,36}-{unix,windows}-{default},pylint
-
-[testenv]
-passenv = HOME
-setenv =
-  SLUGIFY_USES_TEXT_UNIDECODE = yes
-  VIRTUALENV_NO_DOWNLOAD = yes
-deps =
-  -e ../../dagster[test]
-  -e ../../dagster-test
-  -e .[test]
-
-usedevelop = true
-whitelist_externals =
-  /bin/bash
-  echo
-commands =
-  !windows: /bin/bash -c '! pip list --exclude-editable | grep -e dagster -e dagit'
-  echo -e "--- \033[0;32m:pytest: Running tox tests\033[0m"
-  coverage report --omit='.tox/*,**/test_*.py' --skip-covered
-  coverage html --omit='.tox/*,**/test_*.py'
-  coverage xml --omit='.tox/*,**/test_*.py'
-
-[testenv:pylint]
-whitelist_externals =
-  pylint
-basepython = python
-commands =
-  pytest airflow_op_to_dagster_op_tests/test_airflow_op_to_dagster_op.py -k hive
