+++ python_modules/libraries/dagster-airflow/tox.ini
@@ -40,4 +40,4 @@ commands =
 [testenv:claire-test]
 basepython = python
 commands =
-  pytest dagster_airflow_tests/test_airflow_op_to_dagster_op.py -k http
+  pytest dagster_airflow_tests/test_airflow_op_to_dagster_op.py
